refactor: rearrange using groups
move computers, devices, and software to tools directory; add
placeholder pages for blog and portfolio; implement email and FAQ
sections in about page; create now page with current interests and
technologies.
