+++ app/(main)/about/faq.tsx
