+++ app/(tools)/tech/computers.tsx
