+++ app/(tools)/tech/devices.tsx
